backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
django-rest-framework-simplejwt==5.3.0
django-templated-mail==1.1.1
celery==5.3.6
redis==5.0.1
ratelimit==2.2.1
tenacity==8.2.3
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
fake-useragent==1.4.0 